+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+##
+- Add `Accessory` address encoding/decoding
+
 ## 0.39.0
 - Bugfix disable LOGON_ENABLE after LOGON_SELECT ([#38](https://github.com/ZIMO-Elektronik/DCC/issues/38))